# app/swagger.yml
swagger: "2.0"
info:
  description: This is the REST API for Project Artemis
  version: "1.0.0"
  title: Project Artemis REST API
consumes:
  - application/json
produces:
  - application/json
basePath: /api
...
paths:
...
  /user:
    post:
      operationId: user.create
      tags:
        - User
      summary: Add a user to the system
      description: Add a user to the system database
      parameters:
        - name: user
          in: body
          description: User you would like to add
          required: True
          schema:
            type: object
            properties:
              user_email:
                type: string
                description: User's unique email address
              user_name:
                type: string
                description: Public user name of the user
              user_fingerprint:
                type: string
                description: User's browser fingerprint hash
              user_register_ip:
                type: string
                description: IP address of user during registration
      responses:
        201:
          description: User successfully added
          schema:
            properties:
              user_id:
                type: integer
                description: User's ID
  /user/{user_id}:
    get:
      operationId: user.read_one
      tags:
        - User
      summary: Grab a user's profile information
      description: Grab a user's profile information
      parameters:
        - name: user_id
          in: path
          description: User ID of the user
          type: integer
          required: True
      responses:
        200:
          description: Successfully fetched the user's profile
          schema:
            type: object
            properties:
              user_id:
                type: integer
                description: User's ID
              user_email:
                type: string
                description: User's unique email address
              user_name:
                type: string
                description: Public user name of the user
              user_fingerprint:
                type: string
                description: User's browser fingerprint hash
              user_name_display:
                type: string
                description: User's display name
              user_avatar:
                type: string
                description: User's profile avatar
              user_referral_code:
                type: string
                description: User's referral code
              user_wallet_address:
                type: string
                description: User's wallet address
              user_did_kyc:
                type: boolean
                description: Whether the user completed KYC
              user_did_aml:
                type: boolean
                description: Whether the user completed AML
              user_tcreate:
                type: string
                description: Datetime the user registered
              user_tlogin:
                type: string
                description: Datetime of user's last login

    put:
      operationId: user.update
      tags:
        - User
      summary: Update the user's profile
      description: Update the user's profile
      parameters:
        - name: user_id
          in: path
          description: User's system user ID
          type: integer
          required: True
        - name: user
          in: body
          schema:
            type: object
            properties:
              user_fingerprint:
                type: string
                description: User's browser fingerprint hash
              user_name_display:
                type: string
                description: User's display name
              user_photo_id:
                type: string
                description: Private URL to the cloud bucket object containing the user's Government issued photo ID
              user_avatar:
                type: string
                description: User's profile avatar
              user_wallet_address:
                type: string
                description: User's wallet address
              user_did_kyc:
                type: boolean
                description: Whether the user completed KYC
              user_did_aml:
                type: boolean
                description: Whether the user completed AML
              user_wallet_address_eth:
                type: string
                description: The user's Ethereum wallet address, used for confirming deposits externally on the block chain
              user_wallet_address_btc:
                type: string
                description: The user's Bitcoin wallet address, used for confirming deposits externally on the block chain
              user_wallet_address_xmr:
                type: string
                description: The user's Monero wallet address, used for confirming deposits externally on the block chain
              user_register_ip:
                type: string
                description: IP address of the user used when registering
              user_last_ip:
                type: string
                description: Last IP address of the user used when logging in
              user_tlogin:
                type: string
                description: Datetime of user's last login
      responses:
        200:
          description: Successfully updated the user's profile
          schema:
            properties:
              user_id:
                type: integer
                description: User's system user ID

    delete:
      operationId: user.delete
      tags:
        - User
      summary: Remove a user
      description: Remove a user
      parameters:
        - name: user_id
          in: path
          type: integer
          description: ID of the user to remove
          required: true
      responses:
        200:
          description: Successfully removed the user


  # ===========
  # Transaction
  # ===========
  /transactions:
    get:
      operationId: transaction.read_all
      tags:
        - Transaction
      summary: Retrieve a list of transactions in the system
      description: Retrieve all transactions in the system
      parameters:
        - name: transaction_ids
          in: query
          description: Array of transaction IDs, if blank all transactions will be returned
          type: array
          items:
            type: integer
      responses:
        200:
          description: Successfully retrieved transactions
          schema:
            type: array
            items:
              properties:
                transaction_id:
                  type: integer
                  description: System ID of the transaction
                user_id:
                  type: integer
                  description: ID of the user creating the transaction
                pool_id:
                  type: integer
                  description: System ID of the Pool the related to the transaction
                pool_hash:
                  type: string
                  description: System hash of the Pool related to the transaction
                transaction_amount_buy:
                  type: number
                  description: The amount the user committed to the BUY (Crush) action for the pool or the amount the user deposited into the system
                transaction_amount_sell:
                  type: number
                  description: The amount the user committed to the SELL (Crater) action for the pool
                transaction_is_deposit:
                  type: boolean
                  description: Whether the transaction was a deposit into the user's account
                transaction_wallet_to:
                  type: string
                  description: The payee's wallet address, set as null if it is a deposit
                transaction_wallet_from:
                  type: string
                  description: The payer's wallet address and account to debit from, set to null if its a payout from a pool
                transaction_status:
                  type: integer
                  description: Status of the transaction, 0 = the transaction is being processed, usually after the close of a pool, 1 = This is a transaction paying someone out from winning, -1 = This is a transaction being debited from someone losing, 2 = A deposit transaction from a user to the system, 3 = Transaction is cancelled
                transaction_tcreate:
                  type: string
                  description: Datetime the transaction was initiated
                transaction_tmodified:
                  type: string
                  description: Datetime the transaction was modified
                user:
                  type: object
                  properties:
                    user_id:
                      type: string
                      description: User's system user ID
                    user_name:
                      type: string
                      description: User's public system name
                    user_name_display:
                      type: string
                      description: User's display name
                    user_avatar:
                      type: string
                      description: User's profile avatar
                    user_last_ip:
                      type: string
                      description: Last IP address of the user used when logging in
                    user_tcreate:
                      type: string
                      description: Datetime of when the user registered
                    user_status:
                      type: integer
                      description: System status of the user, 0 = User has no confirmed their email, 1 = User has confirmed their email, 2 = User is validated and completed requisite KYC/AM and can trade, 3 = User is suspended from the system temporarily (likely investigation), 4 = User has been banned from the system
                    user_tmodified:
                      type: string
                      description: Datetime of when the user's account was last updated
                    user_tlogin:
                      type: string
                      description: Datetime of when the user last logged into the system
...
  /transaction:
    post:
      operationId: transaction.create
      tags:
        - Transaction
      summary: Create a transaction associated with the user
      description: Create a transaction associated with the user
      parameters:
        - name: transaction
          in: body
          description: Details of the transaction
          required: True
          schema:
            type: object
            properties:
              user_id:
                type: integer
                description: ID of the user creating the transaction
              pool_id:
                type: integer
                description: System ID of the Pool the related to the transaction
              pool_hash:
                type: string
                description: System hash of the Pool related to the transaction
              transaction_amount_buy:
                type: number
                description: The amount the user committed to the BUY (Crush) action for the pool or the amount the user deposited into the system
              transaction_amount_sell:
                type: number
                description: The amount the user committed to the SELL (Crater) action for the pool
              transaction_is_deposit:
                type: boolean
                description: Whether the transaction was a deposit into the user's account
              transaction_wallet_to:
                type: string
                description: The payee's wallet address, set as null if it is a deposit
              transaction_wallet_from:
                type: string
                description: The payer's wallet address and account to debit from, set to null if its a payout from a pool
              transaction_status:
                type: integer
                description: Status of the transaction, 0 = the transaction is being processed, usually after the close of a pool, 1 = This is a transaction paying someone out from winning, -1 = This is a transaction being debited from someone losing, 2 = A deposit transaction from a user to the system, 3 = Transaction is cancelled
              transaction_tcreate:
                type: string
                description: Datetime the transaction was initiated
              transaction_tmodified:
                type: string
                description: Datetime the transaction was modified
      responses:
        201:
          description: Successfully created transaction
          schema:
            properties:
              user_id:
                type: integer
                description: User's system user ID
              transaction_id:
                type: integer
                description: ID of the transaction
              transaction_tcreate:
                type: string
                description: UTC datetime of the transaction
  /transaction/{transaction_id}:
    get:
      operationId: transaction.read_one
      tags:
        - Transaction
      summary: View a particular transaction ID
      description: View a particular transaction ID
      parameters:
        - name: transaction_id
          in: path
          description: Transaction ID
          type: integer
          required: True
      responses:
        200:
          description: Successfully retrieved transaction
          schema:
            type: object
            properties:
              transaction_id:
                type: integer
                description: ID of the transaction
              user_id:
                type: integer
                description: ID of the user creating the transaction
              pool_id:
                type: integer
                description: System ID of the Pool the related to the transaction
              pool_hash:
                type: string
                description: System hash of the Pool related to the transaction
              transaction_amount_buy:
                type: number
                description: The amount the user committed to the BUY (Crush) action for the pool or the amount the user deposited into the system
              transaction_amount_sell:
                type: number
                description: The amount the user committed to the SELL (Crater) action for the pool
              transaction_is_deposit:
                type: boolean
                description: Whether the transaction was a deposit into the users account
              transaction_wallet_to:
                type: string
                description: The payee wallet address, set as null if it is a deposit
              transaction_wallet_from:
                type: string
                description: The payers wallet address and account to debit from, set to null if its a payout from a pool
              transaction_status:
                type: integer
                description: Status of the transaction, 0 = the transaction is being processed, usually after the close of a pool, 1 = This is a transaction paying someone out from winning, -1 = This is a transaction being debited from someone losing, 2 = A deposit transaction from a user to the system, 3 = Transaction is cancelled
              transaction_tcreate:
                type: string
                description: Datetime the transaction was initiated
              transaction_tmodified:
                type: string
                description: Datetime the transaction was modified

    put:
      operationId: transaction.update
      tags:
        - Transaction
      summary: Update a transaction associated with a user
      description: Update a transaction associated with a user
      parameters:
        - name: user_id
          in: path
          description: User's ID
          type: integer
          required: True
        - name: transaction_id
          in: path
          description: Transaction ID
          type: integer
          required: True
        - name: transaction
          in: body
          schema:
            type: object
            properties:
              transaction_status:
                type: integer
                description: Status of the transaction, 0 = the transaction is being processed, usually after the close of a pool, 1 = This is a transaction paying someone out from winning, -1 = This is a transaction being debited from someone losing, 2 = A deposit transaction from a user to the system, 3 = Transaction is cancelled
      responses:
        200:
          description: Successfully updated transaction
          schema:
            properties:
              transaction_id:
                type: string
                description: Transaction's ID
              user_id:
                type: integer
                description: User's ID
              transaction_status:
                type: integer
                description: Status of the transaction, 0 = the transaction is being processed, usually after the close of a pool, 1 = This is a transaction paying someone out from winning, -1 = This is a transaction being debited from someone losing, 2 = A deposit transaction from a user to the system, 3 = Transaction is cancelled

    delete:
      operationId: transaction.delete
      tags:
        - Transaction
      summary: Delete a user's transaction
      description: Delete a user's transaction
      parameters:
        - name: transaction_id
          in: path
          description: Transaction ID
          type: integer
          required: True
      responses:
        200:
          description: Successfully deleted a transaction


  # ====
  # Pool
  # ====
...